==> 02_PhylogeneticTreeBuilding.Rmd <==
---
title: "02_PhyogeneticTreeBuilding"
author: "Katie Heineman"
date: "6/17/2019"
output: html_document
---

# Set Up
```{r dataimport}
library(data.table)
library(stringr)
library(readr)
# Phylomaker code will not work without this package 
library(phytools)
library(R.utils)
library(ape)
#devtools::install_github("jinyizju/V.PhyloMaker")
library(V.PhyloMaker)
# package for adding subspecies ot species
library(pastis)
library(ips)
library(ggtree)
library(DT)
library(lubridate)
library(httr)
library(RCurl)
library(rjson)
library(jsonlite)

# Reading in Caspio Token
source("Tokens/caspioSource.R")
# Reading in caspio GET command (don't know if this works for things > 1000)
source("caspioFunctions.R")

# Capitlize first letter of string
firstup <- function(x) {
  substr(x, 1, 1) <- toupper(substr(x, 1, 1))
  x
}


# Reading in CaPR species list from caspio
caprSpp <- caspio_get_table_all("tblSpeciesJepCAPR")

# Reading in CNPS
cnps <- caspio_get_table_all("tblCNPSRanks")

```


## Matching Genera to the V Phylomaker Super Tree

To have as much phylogenetic resolution as possible, we want to the genera from our species list (Jepson) to specify under "genus.relative" the most closely related genera in the Phylomaker Super Tree. I did this manually for about 128 genera and I have indicated for each match the citation key for th reference in my Zotero Library. This tree differs significantly from the super tree used in S Phylomaker. I have another file with that data clean up (which I did first)


```{r v phylo maker data clean up}
###################### V Phylomaker  ########################

# Merging capr with cnps so that we can include species with the following criteria:
Spp1 <-  merge(caprSpp,cnps, by="JepID",all.x=T)
Spp1 <- Spp1[ grepl("1B",CRPR) |  grepl("2B",CRPR)|  grepl("3",CRPR)|  grepl("4",CRPR)| TJM1TJM2%in%c("TJM1","TJM2")]
Spp1 <- Spp1[,.(species=name_minus_authors,familyJep=firstup(tolower(family)), genus, TJM1TJM2,CRPR,major_clade,endemicity)]
Spp1[,genus:=word(`species`,1,1),by="species"]

## Finding Family's from genera from phylomaker Nodes file
genusTable1 = as.data.table(nodes.info.1)[,.(familyTree=family[1]),by="genus"]

## Merge megatree families onto genera
Spp1 <- merge(Spp1,genusTable1,by="genus",all.x=T)
Spp1 <- Spp1[,.(species,genus,family=familyTree,familyJep,major_clade)]

# Which genera weren't matched in the mega tree
GenNa1 <- Spp1[is.na(family),.(Count=.N),by=c("genus","familyJep")]
# Checking which families in the CaPR list are not in the super tree
GenNa1[!(familyJep%in%genusTable1$family)][order(familyJep)]

# Making a few fixes to these famileis
GenNa1[familyJep=="Agavaceae", familyJep:="Asparagaceae"]
GenNa1[familyJep=="Chenopodiaceae", familyJep:="Amaranthaceae"]
#GenNa1[genus=="Aphyllon", familyJep:="Orobanchaceae"]
#GenNa1[genus=="Mimetanthe", familyJep:="Phyrmaceae"]

#GenNa1[!(Family%in%genusTable1$family),Family:="FernOrNonVascular"]


# Assigning Genus Relatives
GenNa1[genus%in%c("Acanthoscyphus","Oxytheca"), `:=` (genus.relative="Sidotheca",citation="Reveal_JohanneshowelliaPolygonaceaeEriogonoideae_2004")]
GenNa1[genus%in%c("Achyrachaena"), `:=` (genus.relative="Blepharipappus",citation="Baldwin_Originrelationshipstarweedsilversword_2000")]
GenNa1[genus%in%c("Achyronychia"), `:=` (genus.relative="Loeflingia",citation="Greenberg_Molecularsystematicscharacter_2011")]
GenNa1[genus%in%c("Aconogonon"), `:=` (genus.relative="Polygonum",citation="ThePlantList")]
GenNa1[genus%in%c("Allotropa"), `:=` (genus.relative="Monotropa",citation="Cullings_MolecularphylogenyMonotropoideae_1994")]
GenNa1[genus%in%c("Ammoselinum","Apiastrum"), `:=` (genus.relative="Spermolepis",citation="Nesom_TaxonomyApiastrumammoselinum_2012")]
GenNa1[genus%in%c("Amphipappus"), `:=` (genus.relative="Chrysothamnus",citation="Roberts_MolecularPhylogenyChrysothamnus_2004")]
GenNa1[genus%in%c("Amsonia"), `:=` (genus.relative="Melodinus",citation="Simoes_PHYLOGENYSYSTEMATICSRAUVOLFIOIDEAE_2007")]
GenNa1[genus%in%c("Ancistrocarphus"), `:=` (genus.relative="Stylocline",citation="Morefield_THREENEWSPECIES_1992")]
GenNa1[genus%in%c("Argyrochosma"), `:=` (genus.relative="Pellaea",citation="Sigel_Speciesrelationshipsfarina_2011")]
GenNa1[genus%in%c("Aristocapsa"), `:=` (genus.relative="Chorizanthe",citation="Reveal_JohanneshowelliaPolygonaceaeEriogonoideae_2004")]
GenNa1[genus%in%c("Aspidotis","Astrolepis"), `:=` (genus.relative="Cheilanthes",citation="Gastony_CheilanthoidFernsPteridaceae_1998")]
GenNa1[genus%in%c("Bahiopsis","Heliomeris"), `:=` (genus.relative="Viguiera",citation="Schilling_TAXONOMICREVISIONVIGUIERA_1990")]
GenNa1[genus%in%c("Bebbia"), `:=` (genus.relative="Tridax",citation="Panero_Phylogeneticrelationshipssubtribe_1999")]
GenNa1[genus%in%c("Botrypus","Sceptridium"), `:=` (genus.relative="Botrychium",citation="Dauphin_Molecularphylogeneticssupports_2014")]
GenNa1[genus%in%c("Callitriche"), `:=` (genus.relative="Hippuris",citation="Albach_Piecingtogethernew_2005")]
GenNa1[genus%in%c("Callitropsis"), `:=` (genus.relative="Cupressus",citation="ThePlantList")]
GenNa1[genus%in%c("Calyptridium"), `:=` (genus.relative="Cistanthe",citation="Hershkovitz_PhylogeneticAssessmentRevised_1991")]
GenNa1[genus%in%c("Camissoniopsis","Chylismia","Chylismiella"), `:=` (genus.relative="Camissonia",citation="Hershkovitz_PhylogeneticAssessmentRevised_1991")]
GenNa1[genus%in%c("Canbya"), `:=` (genus.relative="Argemone",citation="Jork_MolecularphylogenyOld_1995")]
GenNa1[genus%in%c("Carsonia"), `:=` (genus.relative="Cleome",citation="ThePlantList")]
GenNa1[genus%in%c("Centromadia"), `:=` (genus.relative="Hemizonia",citation="Baldwin_NewCombinationsNew_1999")]
GenNa1[genus%in%c("Centrostegia"), `:=` (genus.relative="Chorizanthe",citation="Reveal_JohanneshowelliaPolygonaceaeEriogonoideae_2004")]
GenNa1[genus%in%c("Chamerion"), `:=` (genus.relative="Epilobium",citation="Levin_ParaphylytribeOnagreae_2004")]
GenNa1[genus%in%c("Chloropyron","Dicranostegia"), `:=` (genus.relative="Cordylanthus",citation="Tank_Phylogeneticclassificationsubtribe_2009")]
GenNa1[genus%in%c("Coleogyne"), `:=` (genus.relative="Kerria",citation="Potter_PhylogenyclassificationRosaceae_2007")]
GenNa1[genus%in%c("Extriplex","Stutzia"), `:=` (genus.relative="Atriplex",citation="Kadereit_MolecularphylogenyAtripliceae_2010")]
GenNa1[genus%in%c("Monolepis"), `:=` (genus.relative="Spinacia",citation="Kadereit_MolecularphylogenyAtripliceae_2010")]
GenNa1[genus%in%c("Orogenia"), `:=` (genus.relative="Lomatium",citation="Sun_ITSBasedPhylogeneticAnalysis_2004a")]
GenNa1[genus%in%c("Sphenosciadium"), `:=` (genus.relative="Angelica",citation="Sun_ITSBasedPhylogeneticAnalysis_2004a")]
GenNa1[genus%in%c("Pilostyles"), `:=` (genus.relative="Apodanthes",citation="Blarer_Comparativefloralstructure_2004")]
GenNa1[genus%in%c("Landoltia"), `:=` (genus.relative="Spirodela",citation="Tippery_Evaluationphylogeneticrelationships_2015")]
GenNa1[genus%in%c("Dicoria","Iva"), `:=` (genus.relative="Xanthium",citation="Miao_SystematicsImplicationsChloroplast_1995")]
GenNa1[genus%in%c("Hymenopappus"), `:=` (genus.relative="Marshallia",citation="Karis_MorphologicalphylogeneticsAsteraceaeAsteroideae_1993")]
GenNa1[genus%in%c("Jensia"), `:=` (genus.relative="Madia",citation="Baldwin_Originrelationshipstarweedsilversword_2000")]
GenNa1[genus%in%c("Malperia"), `:=` (genus.relative="Hofmeisteria",citation="ThePlantList")]
GenNa1[genus%in%c("Phalacroseris"), `:=` (genus.relative="Krigia",citation="Lee_PhylogeneticRelationshipsPrimarily_2003")]
GenNa1[genus%in%c("Rudbeckia"), `:=` (genus.relative="Helianthus",citation="Urbatsch_PhylogeneticAffinitiesConeflower_1995")]
GenNa1[genus%in%c("Sanvitalia"), `:=` (genus.relative="Zaluzania",citation="Panero_Phylogeneticrelationshipssubtribe_1999")]
GenNa1[genus%in%c("Stebbinsoseris"), `:=` (genus.relative="Microseris",citation="Panero_Phylogeneticrelationshipssubtribe_1999")]
GenNa1[genus%in%c("Verbesina"), `:=` (genus.relative="Encelia",citation="Panero_Chloroplastdnarestriction_1997")]
GenNa1[genus%in%c("Struthiopteris","Woodwardia"), `:=` (genus.relative="Sadleria",citation="deGasper_Molecularphylogenyfern_2017")]
GenNa1[genus%in%c("Eucrypta"), `:=` (genus.relative="Phacelia",citation="Ferguson_PhylogeneticAnalysisRelationships_1998")]
GenNa1[genus%in%c("Howellanthus"), `:=` (genus.relative="Phacelia",citation="Gilbert_PhylogeneticRelationshipsPhacelia_2005")]
GenNa1[genus%in%c("Idahoa","Subularia"), `:=` (genus.relative="Xerodraba",citation="Warwick_Closinggapsphylogenetic_2010")]
GenNa1[genus%in%c("Transberingia"), `:=` (genus.relative="Crucihimalaya",citation="Bailey_GlobalPhylogenyBrassicaceae_2006")]
GenNa1[genus%in%c("Grusonia"), `:=` (genus.relative="Corynopuntia",citation="Griffith_PhylogenyOpuntioideaeCactaceae_2009")]
GenNa1[genus%in%c("Githopsis"), `:=` (genus.relative="Heterocodon",citation="Zhou_Phylogeneticassessmentbiogeographic_2012")]
GenNa1[genus%in%c("Howellia","Porterella","Legenere"), `:=` (genus.relative="Downingia",citation="Schultheis_SystematicsDowningiaCampanulaceae_2001")]
GenNa1[genus%in%c("Nemacladus"), `:=` (genus.relative="Campanula",citation="Cosner_PhylogeneticrelationshipsCampanulales_1994")]
GenNa1[genus%in%c("Triodanis"), `:=` (genus.relative="Legousia",citation="Cosner_ChloroplastDNArearrangements_2004")]
GenNa1[genus%in%c("Scopulophila"), `:=` (genus.relative="Drymaria",citation="Smissen_SubfamilialrelationshipsCaryophyllaceae_2002")]
GenNa1[genus%in%c("Crocanthemum"), `:=` (genus.relative="Cistus",citation="Aparicio_Phylogeneticreconstructiongenus_2017")]
GenNa1[genus%in%c("Oxystylis","Peritoma","Wislizenia"), `:=` (genus.relative="Cleomella",citation="Riser_PhylogeneticrelationshipsNorth_2013")]
GenNa1[genus%in%c("Sedella"), `:=` (genus.relative="Sedum",citation="Nikulin_PhylogeneticrelationshipsSedum_2016")]
GenNa1[genus%in%c("Hesperocyparis"), `:=` (genus.relative="Cupressus",citation="ThePlantList")]
GenNa1[genus%in%c("Hemitomes"), `:=` (genus.relative="Monotropa",citation="Cullings_Singlephylogeneticorigin_1996")]
GenNa1[genus%in%c("Pityopus","Pterospora"), `:=` (genus.relative="Pleuricospora",citation="Cullings_Singlephylogeneticorigin_1996")]
GenNa1[genus%in%c("Sarcodes"), `:=` (genus.relative="Monotropa",citation="Kron_PhylogeneticRelationshipsEmpetraceae_1996")]
GenNa1[genus%in%c("Hosackia"), `:=` (genus.relative="Lotus",citation="Allan_Tribaldelimitationphylogenetic_2000")]
GenNa1[genus%in%c("Pickeringia"), `:=` (genus.relative="Styphnolobium",citation="Wojciechowski_PhylogenyLegumesLeguminosae_2004")]
GenNa1[genus%in%c("Notholithocarpus"), `:=` (genus.relative="Quercus",citation="Manos_Phylogeneticrelationshipstaxonomic_2008")]
GenNa1[genus%in%c("Comastoma"), `:=` (genus.relative="Gentianella",citation="vonHagen_Phylogenyflowerevolution_2002")]
GenNa1[genus%in%c("Frasera"), `:=` (genus.relative="Swertia",citation="vonHagen_Phylogenyflowerevolution_2002")]
GenNa1[genus%in%c("Eustoma","Zeltnera"), `:=` (genus.relative="Sabatia",citation="Mansion_Genericdelimitationphylogenetic_2004")]
GenNa1[genus%in%c("Condea"), `:=` (genus.relative="Hyptis",citation="Mansion_Genericdelimitationphylogenetic_2004")]
GenNa1[genus%in%c("Trichostema"), `:=` (genus.relative="Caryopteris",citation="Wagstaff_PhylogenyLabiataeinferred_1998")]
GenNa1[genus%in%c("Streptopus"), `:=` (genus.relative="Prosartes",citation="Fay_PhylogeneticsLilliales_2006")]
GenNa1[genus%in%c("Floerkea"), `:=` (genus.relative="Limnanthes",citation="Meyers_molecularphylogenyLimnanthes_2010")]
GenNa1[genus%in%c("Herissantia"), `:=` (genus.relative="Abutilon",citation="Tate_Phylogeneticrelationshipstribe_2005")]
GenNa1[genus%in%c("Iliamna","Malacothamnus"), `:=` (genus.relative="Alcea",citation="LaDuke_ChloroplastDNABased_1995")]
GenNa1[genus%in%c("Sidalcea"), `:=` (genus.relative="Eremalche",citation="Tate_Phylogeneticrelationshipstribe_2005")]
GenNa1[genus%in%c("Tripterocalyx"), `:=` (genus.relative="Abronia",citation="Douglas_MolecularphylogenyNyctaginaceae_2007")]
GenNa1[genus%in%c("Eremothera","Eulobus"), `:=` (genus.relative="Camissonia",citation="Levin_ParaphylytribeOnagreae_2004")]
GenNa1[genus%in%c("Neoholmgrenia","Tetrapteron"), `:=` (genus.relative="Camissonia",citation="Allen_PhylogeneticModelsLinking_2017")]
GenNa1[genus%in%c("Taraxia"), `:=` (genus.relative="Gayophytum",citation="Allen_PhylogeneticModelsLinking_2017")]
GenNa1[genus%in%c("Listera"), `:=` (genus.relative="Vanda",citation="Neyland_PhylogenySubfamilyEpidendroideae_1996")]
GenNa1[genus%in%c("Piperia"), `:=` (genus.relative="Platanthera",citation="Bateman_Molecularphylogeneticsmorphological_2009")]
GenNa1[genus%in%c("Kopsiopsis"), `:=` (genus.relative="Boschniakia",citation="McNeal_Phylogenyoriginsholoparasitism_2013")]
GenNa1[genus%in%c("Meconella"), `:=` (genus.relative="Roemeria",citation="Kadereit_phylogenyPapaverPapaveraceae_1997")]
GenNa1[genus%in%c("Gambelia"), `:=` (genus.relative="Galvezia",citation="Ogutcen_phylogeneticstudytribe_2016")]
GenNa1[genus%in%c("Holmgrenanthe"), `:=` (genus.relative="Asarina",citation="Ghebrehiwet_PhylogenytribeAntirrhineae_2000")]
GenNa1[genus%in%c("Lindernia"), `:=` (genus.relative="Scrophularia",citation="Estes_Phylogeneticrelationshipsmonotypic_2008")]
GenNa1[genus%in%c("Maurandya"), `:=` (genus.relative="Maurandya",citation="Ogutcen_phylogeneticstudytribe_2016")]
GenNa1[genus%in%c("Nuttallanthus"), `:=` (genus.relative="Linaria",citation="Crawford_Geneticvariationreproductive_2006")]
GenNa1[genus%in%c("Pseudorontium"), `:=` (genus.relative="Antirrhinum",citation="Keil_MOHAVEASUBSUMEDANTIRRHINUM_2018")]
GenNa1[genus%in%c("Tonella"), `:=` (genus.relative="Collinsia",citation="Baldwin_Phylogeneticperspectivesdiversification_2011")]
GenNa1[genus%in%c("Dasyochloa"), `:=` (genus.relative="Munroa",citation="Columbus_PhylogeneticsChloridoideaeGramineae_2007")]
GenNa1[genus%in%c("Dissanthelium"), `:=` (genus.relative="Poa",citation="Columbus_PhylogeneticsChloridoideaeGramineae_2007")]
GenNa1[genus%in%c("Scribneria"), `:=` (genus.relative="Deschampsia",citation="Soreng_worldwidephylogeneticclassification_2017")]
GenNa1[genus%in%c("Torreyochloa"), `:=` (genus.relative="Deschampsia",citation="Soreng_phylogeneticanalyischloroplast_1990")]
GenNa1[genus%in%c("Vahlodea"), `:=` (genus.relative="Deschampsia",citation="Chiapella_MolecularPhylogeneticStudy_2007")]
GenNa1[genus%in%c("Aconogonon"), `:=` (genus.relative="Koenigia",citation="Galasso_MolecularphylogenyPolygonum_2009")]
GenNa1[genus%in%c("Mucronea"), `:=` (genus.relative="Chorizanthe",citation="Kempton_SystematicsEriogonoideaePolygonaceae_2012")]
GenNa1[genus%in%c("Nemacaulis"), `:=` (genus.relative="Chorizanthe",citation="Kempton_SystematicsEriogonoideaePolygonaceae_2012")]
GenNa1[genus%in%c("Rumex"), `:=` (genus.relative="Rheum",citation="Schuster_updatedmolecularphylogeny_2015")]
GenNa1[genus%in%c("Systenotheca"), `:=` (genus.relative="Chorizanthe",citation="Kempton_SystematicsEriogonoideaePolygonaceae_2012")]
GenNa1[genus%in%c("Myriopteris"), `:=` (genus.relative="Cheilanthes",citation="Grusz_monophyleticCheilanthesresurrection_2013")]
GenNa1[genus%in%c("Notholaena"), `:=` (genus.relative="Cheilanthes",citation="Rothfels_MonophyleticNotholaenaPteridaceae_2008")]
GenNa1[genus%in%c("Pentagramma"), `:=` (genus.relative="Pityrogramma",citation="Schuettpelz_unifiedapproachtaxonomic_2015")]
GenNa1[genus%in%c("Poteridium"), `:=` (genus.relative="Sanguisorba",citation="Potter_PhylogenyclassificationRosaceae_2007")]
GenNa1[genus%in%c("Hemieva"), `:=` (genus.relative="Suksdorfia",citation="Soltis_KaryotypesRelationshipsBolandra_1987")]
GenNa1[genus%in%c("Mitellastra"), `:=` (genus.relative="Elmera",citation="Okuyama_Parallelfloraladaptations_2008")]
GenNa1[genus%in%c("Ozomelis"), `:=` (genus.relative="Conimitella",citation="Folk_Phylogeneticrelationshipscharacter_2014")]
GenNa1[genus%in%c("Pectiantia"), `:=` (genus.relative="Mitella",citation="Folk_Phylogeneticrelationshipscharacter_2014")]
GenNa1[genus%in%c("Euphrosyne"), `:=` (genus.relative="Ambrosia",citation="Miao_SystematicsImplicationsChloroplast_1995")]
GenNa1[genus%in%c("Spartina"), `:=` (genus.relative="Sporobolus",citation="Hilu_phylogenyChloridoideaePoaceae_2001")]
GenNa1[genus%in%c("Aphyllon"), `:=` (genus.relative="Orobanche",citation="Schneider_ResurrectiongenusAphyllon_2016")]
GenNa1[genus%in%c("Mimetanthe"), `:=` (genus.relative="Mimulus",citation="XXX")]
GenNa1[genus%in%c("Hesperoyucca"), `:=` (genus.relative="Yucca",citation="XXX")]

# Update familes in the mega tree list
setkey(Spp1, genus)
setkey(GenNa1, genus)
Spp1[GenNa1,`:=` (genus = i.genus,family=i.familyJep, genus.relative=i.genus.relative)]

# Adding species relatives to varieties and subspecies
Spp1[grepl("var.",species, fixed=T) | grepl("subsp.",species, fixed=T),species.relative:=word(species,1,2)]



```

# Looking at how V Phylomaker Binds Species with Abies as example genus

I specified that Abies magnifica var. magnifica and var. shastensis are relatives of Abies magnifica. This adds the subspecies to the same polytomy as Abies magnifica, but it does not nest them within Abies magnifica

```{r abies plot}
# Binding Relatives
abiesBound <- bind.relative(SppVas1[genus=="Abies"]) 
# Building Tree
result_california_treeV_abies<- phylo.maker(sp.list=abiesBound$species.list, tree=abiesBound$phylo, nodes=abiesBound$nodes.info) 

plot(result_california_treeV_abies$scenario.3)

```

# Function for Adding InfraSpecies Nested within Species

What I would rather see is susbspecies and varieties replaces a species on a phylogeny. I wrote a regular expression code to replace species taxon with nested infra taxa. I make a somewhat large assumption in dealing with branch lengths: For each group of infraspecies, I added a poltomy at the halfway point of the species existing branch. Then the branch length for each infraspecies is equal to half of the branch length of the original species. I made this assumption to make sure I didn't make infraspecies have artificially longer branches than non-infraspecies. 


For instance, for a clade in which _Abies procera_ and _Abies magnifica_ are sister species [see Old Tree], I added two subspecies for _Abies magnifica_ ( _Abies magnifica var.magnifica _ & _Abies magnifica var. sharstensis _), by replacing _Abies mangifica_ with a string that has two infra species with equal branch lengths that diverge at a node half way down original _Abies magnifica_ branch

**OLD TREE**
(Abies_procera:0.58,Abies_magnifica:0.58)


**NEW TREE**
(Abies_procera:0.58,(Abies_magnifica_var_magnifica:0.29, Abies_magnifica_var_shastensis:0.29):0.29)


```{r abies example}

addSubSpecies <- function(infraList,Tree){
  # Find the species you want to replace with infra species
  species <- word(infraList[1],1,2,sep="_")
  # Making the tree a sring
  TreeText <- write.tree(Tree)
    
  # Extract the branch length next to the species
  branchLength <- as.numeric(gsub(paste0("^.*",species,".*?(\\d+\\.*\\d*).*"),"\\1", TreeText))

  # Calculate new species branch length 
  # HUGE ASSUMPTION HERE THAT BRANCH LENGTH FOR INFRASPECIES IS HALF THAT OF BRANCH LENGTHS BETWEEN SPECIES
  createSppBL <- function(infra){
    paste0(infra,":",branchLength/2)
  }

# Creating String of subspecies that will replace "species"
 # subSpeciesString <- paste0("(",toString(sapply(infraList,createSppBL)),"):")
  subSpeciesString <- paste0("(",toString(sapply(infraList,createSppBL)),"):",as.character(branchLength/2))

  #print(subSpeciesString)
  # Replacing species with subspecies
  #FixedTree <- read.newick(text=gsub(paste0(species,":"),subSpeciesString,TreeText))
  FixedTree <- read.newick(text=gsub(paste0(species,":\\d+\\.*\\d*"),subSpeciesString,TreeText))
  
  return(FixedTree)
}

## Abies Example
# Abies phylogeny only species level taxa
AbiesNoSub <- phylo.maker(sp.list=SppVas1[genus=="Abies" & !(grepl("var.",species))]) 

plot(AbiesNoSub$scenario.3)

# List of subspecies I want to add
infraListAbies <- c("Abies_magnifica_var._magnifica","Abies_magnifica_var._shastensis")

# Running tree and code through list
newAbiesTree <- addSubSpecies(infraListAbies,AbiesNoSub$scenario.3)

plot(newAbiesTree)

## Astragalus example
AstraNoSub <- phylo.maker(sp.list=SppVas1[genus=="Astragalus" & !(grepl("var.",species,fixed=T) | grepl("subsp.",species,fixed=T))]) 

# List of subspecies I want to add
infraListAstrLen <- SppVas1[grepl("lentiginosus",species) & (grepl("var.",species,fixed=T) | grepl("subsp.",species,fixed=T))]$species
infraListAstrLen <- gsub(" ","_",infraListAstrLen)

# Running tree and code through list
newAstrTree <- addSubSpecies(infraListAstrLen,AstraNoSub$scenario.3)

plot(newAstrTree)

```

# Assembling the Phylogenetic Tree with V Phylomaker

V.Phylomater produces phylogenies using three scenarios. This function & method were described in Qian & Jin 2019 in Journal of Ecology.

1. adding genera or species as basal polytomies within their families or genera 
2. adding genera or species randomly within their families or genera 
3. adding genera or species to their families or genera using the same approach implemented in Phylomatic and BLADJ  <- This is deafult

I plan to build a phylogeny for all species level taxa using phylomaker and then nest the subspecies within each species (replacing the taxa levelspecies) 

```{r vphylomaker}
## Bulding Tree for Scenario 3

# sourcing V Phylomaker code
# sapply( list.files("VPhylomakerCodes/", full.names=TRUE), source )

# Filtering for only species
SppCopy <- Spp1
SppCopy[,binomial:=word(species,1,2)]
SppCopy[grepl(" x ", species)|grepl(" X ", species),binomial:=ifelse(str_count(species," ")==3,binomial,word(species,1,3,sep=" "))]
SppCopy[,SppCount:=.N,by="binomial"]
SppCopy[,SppID:=1:.N,by="binomial"]
SppOnly <- SppCopy[SppID==1]
SppOnly[,species:=binomial]
SppOnly[,major_clade:=NULL]

# Binding Relatives at the Genus Level
SpeicesOnlyBind <- bind.relative(SppOnly)

# Running V phylomaker
result_california_speciesOnly <-  phylo.maker(sp.list=SpeicesOnlyBind$species.list, tree=SpeicesOnlyBind$phylo, nodes=SpeicesOnlyBind$nodes.info) 

# Writing data to the disc
  write.tree(result_california_speciesOnly$scenario.3,"Data/result_california_speciesOnly_20190923.tre")
  write.csv(result_california_speciesOnly$species.list,"Data/result_california_speciesOnly_201909123.csv")

```

# Adding infraspecies to the tree

Looping through all the species with infra species and replace the species level taxa with multiple infraspecies in the same clade.

```{r looping through infraspecies code}

#########
# Trying out new method without reinventing wheel
SppCopy[,infraSpp := toString(species[(grepl("var.",species,fixed=T)|grepl("subsp.",species,fixed=T)|grepl(" x ",species,fixed=T)|grepl(" X ",species,fixed=T))]),by="binomial"]

# Asking which collected species have infra species but also binomial level names
SppCopy[, Reassign:=ifelse((binomial==species & SppCount>1 ), "Yes","No")]
SppCopy[Reassign=="Yes",.(binomial,infraSpp)]
SppCopy[grepl("var.",infraSpp,fixed=T),InfraType:="var."]
SppCopy[grepl("subsp.",infraSpp,fixed=T),InfraType:="subsp."]
SppCopy[grepl(" x ",infraSpp,fixed=T)|grepl(" X ",infraSpp,fixed=T),InfraType:="x"]

# Ensuring sub types don't exist (STILL NEED TO CHECK)
SppCopy[grepl("subsp.",infraSpp,fixed=T) & grepl("var.",infraSpp,fixed=T),InfraType:="both"]
SppCopy[grepl(" x ",infraSpp,fixed=T) & (grepl("var.",infraSpp,fixed=T)|grepl("subsp.", infraSpp,fixed=T)),InfraType:="both"]

# Specifying the name on the phylogeny for each species (different than the Jep ID)
SppCopy[,nameOnPhylogeny:=ifelse(Reassign=="Yes",paste(species,InfraType,word(species,2,2),sep=" "),species)]

# Replace spaces with "_"
SppCopy[,nameOnPhylogeny:=gsub(" ","_",nameOnPhylogeny)]

write.csv(SppCopy,"Data/spp_withNameOnPhylogeny_20190923.csv")

# Create object that can be looped through to add species
InfraSummary <- SppCopy[(infraSpp)!="",.(InfraString=toString(unique(nameOnPhylogeny))),by=c("binomial")]

# Looping through the tree
newTree <- result_california_speciesOnly$scenario.3
for (i in 1:length(InfraSummary$binomial)){
#for (i in 1:10){
infraListtmp <- unlist(strsplit(InfraSummary$InfraString[i], split=", "))
print(infraListtmp)
newTree <- addSubSpecies(infraListtmp,newTree)

}

write.tree(newTree, "Data/californiaTreewithInfraSpecies_20190923.tre")

# Evaluating species genera

pruned.treeAbronia<-drop.tip(newTree,newTree$tip.label[-which(grepl("Abronia",newTree$tip.label))])
plot(pruned.treeAbronia)

pruned.treeAbies<-drop.tip(newTree,newTree$tip.label[-which(grepl("Abies",newTree$tip.label))])
plot(pruned.treeAbies)

pruned.treeDud<-drop.tip(newTree,newTree$tip.label[-which(grepl("Dudleya",newTree$tip.label))])
plot(pruned.treeDud)



```
