==> 03_Manipulating_Phylogenetic_Tree.Rmd <==
---
title: "03 Manipulating Phylogenetic Tree"
author: "Katie Heineman"
date: "6/21/2019"
output: html_document
---

# Read in Data & Tree Created in File 02_PhylogeneticTreeBuilding

```{r setup, include=FALSE}
library(ape)
library(data.table)
library(phytools)
library(stringr)
library(caper)
library(picante)


# Species ONLY california phylogeny created in 02 code
sppOnlyTree <- read.tree("Data/result_california_speciesOnly_20190923.tre")
# California phylogeny with infraspecies created in 02 code
infraTree <- read.tree("Data/californiaTreewithInfraSpecies_20190923.tre")
# Accessions Data
accDat <- fread("Data/tblAccessionsCAPR_2019-Jun-21_1829.csv",na.strings=c("",NA,"NA"))
accDat[,organismQuantity:=as.numeric(organismQuantity)]
accDat <- accDat[Deaccession==0]

# Generated from script in 02
spp <- fread("Data/spp_withNameOnPhylogeny_20190626.csv")
# Species tables from DB
allJepSpecies <- fread("Data/tblSpeciesJepCAPR_2019-Jun-27_1654.csv")
cnps <- fread("Data/tblCNPSRanks_2019-Jun-21_1830.csv")

# Looking at the tip.labels that aren't in the spp list
tipLablesInfraTree <- as.data.table(infraTree$tip.label)
spp[!(nameOnPhylogeny%in%tipLablesInfraTree$V1)]

# Need to connect name on phylogeny to Jep ID so that we can get the CNPS ranks
sppJep <- merge(spp, allJepSpecies[,.(JepID,name_minus_authors,endemicity)],by.x="species",by.y="name_minus_authors",all.x=T)
sppJep <- merge(sppJep,cnps, by="JepID",all=T)

# Merge summary onto species and CNPS
AccwithSpp <- merge(sppJep,accDat,by.y="taxonID",by.x="JepID",all.x=T)


```

# Data Summary
```{r summary}
## Summarize Accession Data by JepID to find the following variables
# Total Seed Count
# Total Seed Collections
# Total Accessions
# Total unique Cnddb EO indices


summAcc <- AccwithSpp[,
                      .(collectionTypes=toString(sort(unique(basisofRecord))),
                        totalSeed=sum((organismQuantity),na.rm=T),
                        totalSeedCollections=sum(1*(grepl("Seed",basisofRecord))),
                        totalLivingCollections=sum(1*(grepl("Living",basisofRecord))),
                        totalColl=sum(1*(!is.na(eventID))), 
                        uniqueEOs = toString(unique(cnddbEOIndex)),
                        family=family[1],
                        genus=genus[1],
                        institutions = toString(sort(unique(institutionCode))),
                        CRPR=CRPR[1],
                        Counties=Counties[1]),
                        by="nameOnPhylogeny"]
summAcc[,uniqueEOs:=ifelse(uniqueEOs=="NA", 0,str_count(uniqueEOs,",")+1)]

summAcc[!(nameOnPhylogeny%in%infraTree$tip.label),.(nameOnPhylogeny,family,genus)][order(nameOnPhylogeny)]


```


# Match Traits to Tree

```{r matching traits to tree}

# Making a matrix
traitDT <- summAcc[!is.na(nameOnPhylogeny),.(nameOnPhylogeny,collectionTypes,totalSeed,totalSeedCollections,totalLivingCollections,totalColl,family,genus,institutions,CRPR,Counties)]
#traitDT[is.na(totalSeedCollections),totalSeedCollections:=0]
# YesNo Variables
traitDT[,AnyCollection:=ifelse(totalColl==0,0,1)]
traitDT[,InSeedCollection:=ifelse(totalSeedCollections==0,0,1)]
traitDT[,InLivingCollection:=ifelse(totalLivingCollections==0,0,1)]

# Makea dataframe with row labels
traitDF<- data.frame(traitDT,row.names = traitDT$nameOnPhylogeny)

# Ensuring trait matrix order matches the phylogenetic tree order 
traitDF <- traitDF[infraTree$tip.label,]




# Creating a data object that matches the tree and the traits (only including T50 and GP)
MatchDataObj= match.phylo.data(infraTree, traitDF)

# Calcalculate evoluationary distinctiveness for each species
EvDist <- evol.distinct(MatchDataObj$phy, type = c("equal.splits", "fair.proportion"),
              scale = FALSE, use.branch.lengths = TRUE)
MatchDataObj$data$evolDist <- EvDist$w

# Write to the App Data folder
 saveRDS(MatchDataObj, "caprPrioritizationApp/AppData/MatchDataObj.rds")

# Example of how to filter family
pruned.treePhyrmaceae<-drop.tip(MatchDataObj$phy,MatchDataObj$phy$tip.label[-which(grepl("Phrymaceae",as.character(MatchDataObj$data$family)))])

# Filter by family
viewFamily <- function(family){
  pruned.tree<-drop.tip(MatchDataObj$phy,MatchDataObj$phy$tip.label[-which(grepl(family,as.character(MatchDataObj$data$family)))])
  plot(pruned.tree)
}


# Testing for phylogenetic signal in binary trait
 objPhySeedCOll <- phylo.d(MatchDataObj$data, MatchDataObj$phy, names.col=nameOnPhylogeny, binvar=SeedCollYN, permut = 1000)


# Look at Grasses
viewFamily("Poaceae")

# Look at Oaks
viewFamily("Fagaceae")

```

# Filtering by Traits
